Implement support for NOEXEC
This is using seccomp syscall filtering of the execve and execveat
syscalls. Unlike the original sudo this does not depend on using
LD_PRELOAD. Instead SECCOMP_RET_USER_NOTIF is used combined with a
thread which allows the first exec (as done by sudo itself) and denies
all further execs. This makes it robust against both statically linked
executables and executables compiled for a different libc than sudo-rs
itself. Sudo-rs just like original sudo does not protect against
malicious programs with NOEXEC. It doesn't prevent mapping memory as
executable, nor does it protect against future syscalls that do an exec
like the planned io_uring exec opcode [^1]. And it also doesn't protect
against honest programs that intentionally or not allow the user to
write to /proc/self/mem for the same reasons as that it doesn't protect
against malicious programs.

[^1]: https://lwn.net/Articles/1002371/